requirements_vps.txt
# VPS用 依存パッケージ

# FastAPI + ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LLM Providers
openai>=1.50.0
google-generativeai>=0.8.0
anthropic>=0.18.0

# LINE Bot SDK
line-bot-sdk==3.6.0

# 環境変数管理
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# データベース（SQLite標準ライブラリを使用）
# PostgreSQL
psycopg2-binary==2.9.9

# ログ出力（標準ライブラリを使用）
